fix: some refactoring based on PR comments
